update(Comment, Category, File, Image, Like, Post, User): move entity files
