+++ src/main/java/com/february/edsc/domain/post/comment/Comment.java
@@ -1,9 +1,11 @@
-package com.february.edsc.domain;
+package com.february.edsc.domain.post.comment;
 
+import com.february.edsc.domain.post.Post;
 import lombok.Getter;
 import lombok.Setter;
 
 import javax.persistence.*;
+import java.sql.Timestamp;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -12,8 +14,7 @@ import java.util.List;
 public class Comment {
 
     @Id
-    @GeneratedValue
-    @Column(name = "comment_id")
+    @GeneratedValue(strategy= GenerationType.IDENTITY)
     private Long id;
 
     private String content;
@@ -23,6 +24,12 @@ public class Comment {
 
     @Column(name = "is_public")
     private boolean isPublic;
+
+    @Column(name = "created_at")
+    private Timestamp createdAt;
+
+    @Column(name = "modified_at")
+    private Timestamp modifiedAt;
 
     @ManyToOne(fetch = FetchType.LAZY)
     private Post post;
