+++ src/main/java/com/february/edsc/domain/post/file/File.java
@@ -1,5 +1,6 @@
-package com.february.edsc.domain;
+package com.february.edsc.domain.post.file;
 
+import com.february.edsc.domain.post.Post;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -7,11 +8,10 @@ import javax.persistence.*;
 
 @Entity
 @Getter @Setter
-public class Image {
+public class File {
 
     @Id
-    @GeneratedValue
-    @Column(name = "image_id")
+    @GeneratedValue(strategy= GenerationType.IDENTITY)
     private Long id;
 
     private String path;
